Add new user_dbs queue for jobs
