+++ spec/lib/carto/ghost_tables_manager_spec.rb
@@ -36,7 +36,7 @@ module Carto
       @user.tables.count.should eq 0
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
 
-      ::Resque.expects(:enqueue).with(::Resque::UserJobs::SyncTables::LinkGhostTables, @user.id).never
+      ::Resque.expects(:enqueue).with(::Resque::UserDBJobs::UserDBMaintenance::LinkGhostTables, @user.id).never
 
       @ghost_tables_manager.link_ghost_tables_synchronously
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_true
@@ -126,7 +126,7 @@ module Carto
       @user.tables.count.should eq 0
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
 
-      ::Resque.expects(:enqueue).with(::Resque::UserJobs::SyncTables::LinkGhostTables, @user.id).never
+      ::Resque.expects(:enqueue).with(::Resque::UserDBJobs::UserDBMaintenance::LinkGhostTables, @user.id).never
 
       @ghost_tables_manager.link_ghost_tables_synchronously
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_true
@@ -171,7 +171,7 @@ module Carto
       @user.tables.count.should eq 0
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_false
 
-      ::Resque.expects(:enqueue).with(::Resque::UserJobs::SyncTables::LinkGhostTables, @user.id).never
+      ::Resque.expects(:enqueue).with(::Resque::UserDBJobs::UserDBMaintenance::LinkGhostTables, @user.id).never
 
       @ghost_tables_manager.link_ghost_tables_synchronously
       @ghost_tables_manager.instance_eval { user_tables_synced_with_db? }.should be_true
